requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
auth==0.5.3
authentication==1.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.3
falcon==3.1.1
Flask==2.2.3
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
passlib==1.7.4
pycryptodome==3.17
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pytz==2023.3
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
